refactor: change name of AuthTextInput to FormInput
