+++ src/pages/auth/signUp/index.tsx
@@ -3,7 +3,7 @@ import { getEmailDuplicateCheck } from "@/api/auth/email/check";
 import { SignUpUser, signUpUserAtom } from "@/atoms/auth/signUpUser";
 import AuthButton from "@/components/auth/AuthButton";
 import AuthLayout from "@/components/auth/AuthLayout";
-import AuthTextInput from "@/components/auth/AuthTextInput";
+import AuthTextInput from "@/components/FormInput";
 import KakaoLoginButton from "@/components/auth/KakaoLoginButton";
 import Logo from "@/components/Logo";
 import VALIDATION_RULE from "@/constants/auth/VALIDATION_RULE";
